requirements.txt
python-telegram-bot
transformers
duckduckgo-search
bs4
requests
torch
